requirements.txt
line-bot-sdk==1.16.0
Flask==1.0.2
gunicorn==19.9.0
requests==2.20.0
peewee